Move LoginForm to a component from SignInPage. #10
